requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.5
httpx==0.27.2
dashscope>=1.17.0
PyMuPDF>=1.23.0